requirements.txt
discord.py[voice]
openai
numpy
matplotlib
sphinx
sphinx_rtd_theme